fix: warning in compilation
remove warning with function who manage controler

use static cast to cast float to int

+++ src/Raylib/Raylib.cpp
@@ -212,7 +212,7 @@ bool Raylib::isControllerValid(int const &idx, std::string const &ControllerName
 
 std::vector<int> Raylib::getControllerInput(int idx)
 {
-    std::array<int, 4> joystick = {GetGamepadAxisMovement(0, 0), GetGamepadAxisMovement(0, 2), GetGamepadAxisMovement(0, 3), GetGamepadButtonPressed()};
+    std::array<int, 4> joystick = {static_cast<int>(GetGamepadAxisMovement(0, 0)), static_cast<int>(GetGamepadAxisMovement(0, 2)), static_cast<int>(GetGamepadAxisMovement(0, 3)), GetGamepadButtonPressed()};
     std::vector<int> joystickToInput;
 
     for (std::size_t i = 0; i != 3; ++i)
